support .gui programs
When a file has the same name as another with ".gui" at the end, the .gui file is executed with --parent_vnc_addr [addr]. It is expected to connect to the given address and act as a VNC server. Its content will be presented as part of the dirgui interface as the editor/presenter for that file.